requirements.txt
matplotlib
mlxtend
numpy
pandas
Pillow
seaborn
streamlit
